Moved text splitter file and tests
